Multi package lib refactor (#106)
