+++ packages/react-api/package.json
@@ -53,9 +53,7 @@
     "webpack-cli": "^4.5.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.13.9",
-    "@turf/bbox-polygon": "^6.3.0",
-    "@turf/boolean-intersects": "^6.3.0"
+    "@babel/runtime": "^7.13.9"
   },
   "peerDependencies": {
     "@carto/react-core": "^1.0.0-rc.1",
