+++ packages/react-api/src/index.js
@@ -1,12 +1,5 @@
 export { executeSQL } from './api/SQL';
-export { default as useCartoLayerFilterProps } from './filters/useCartoLayerFilterProps';
-
 export { SourceTypes } from './api/SourceTypes';
 
-// internal packages
-export { 
-    FilterTypes as _FilterTypes,
-    filtersToSQL as _filtersToSQL,
-    getApplicableFilters as _getApplicableFilters
- } from './filters/FilterQueryBuilder';
-export { buildFeatureFilter as _buildFeatureFilter } from './filters/Filter';
+export { default as useCartoLayerFilterProps } from './hooks/useCartoLayerFilterProps';
+
