+++ packages/react-core/__tests__/operations/histogram.test.js
@@ -1,5 +1,5 @@
 import { histogram } from '../../src/operations/histogram';
-import { AggregationTypes } from '../../src/widgets/AggregationTypes';
+import { AggregationTypes } from '../../src/operations/aggregation/AggregationTypes';
 
 const VALUES = [1, 2, 2, 3, 3, 3, 4, 4, 5];
