+++ packages/react-core/package.json
@@ -54,5 +54,9 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.13.9"
+  },
+  "peerDependencies": {
+    "@turf/bbox-polygon": "^6.3.0",
+    "@turf/boolean-intersects": "^6.3.0"
   }
 }
