+++ packages/react-core/src/index.js
@@ -10,3 +10,16 @@ export { throttle } from './utils/throttle';
 export { randomString } from './utils/randomString';
 
 export { makeClosedInterval } from './utils/makeClosedInterval';
+
+export { AggregationTypes } from './operations/aggregation/AggregationTypes';
+export { aggregationFunctions } from './operations/aggregation/values';
+export { groupValuesByColumn } from './operations/groupby';
+export { histogram } from './operations/histogram';
+
+export { 
+  FilterTypes as _FilterTypes,
+  filtersToSQL as _filtersToSQL,
+  getApplicableFilters as _getApplicableFilters
+} from './filters/FilterQueryBuilder';
+export { buildFeatureFilter as _buildFeatureFilter } from './filters/Filter';
+export { viewportFeatures } from './filters/viewportFeatures';
