+++ packages/react-core/src/operations/aggregation/AggregationTypes.js
