+++ packages/react-core/src/operations/groupby.js
