+++ packages/react-core/src/operations/histogram.js
