+++ packages/react-widgets/__tests__/models/CategoryModel.test.js
@@ -5,7 +5,7 @@ import {
   buildSqlQueryToGetCategories,
   filterViewportFeaturesToGetCategories
 } from '../../src/models/CategoryModel';
-import { AggregationTypes } from '../../src/widgets/AggregationTypes';
+import { AggregationTypes } from '@carto/react-core';
 import { SourceTypes } from '@carto/react-api';
 
 import { mockSqlApiRequest, mockClear } from '../mockSqlApiRequest';
