+++ packages/react-widgets/__tests__/models/HistogramModel.test.js
@@ -5,7 +5,7 @@ import {
   buildSqlQueryToGetHistogram,
   filterViewportFeaturesToGetHistogram
 } from '../../src/models/HistogramModel';
-import { AggregationTypes } from '../../src/widgets/AggregationTypes';
+import { AggregationTypes } from '@carto/react-core';
 import { SourceTypes } from '@carto/react-api';
 
 import { mockSqlApiRequest, mockClear } from '../mockSqlApiRequest';
