+++ packages/react-widgets/package.json
@@ -58,6 +58,7 @@
   },
   "peerDependencies": {
     "@carto/react-api": "^1.0.0-rc.1",
+    "@carto/react-core": "^1.0.0-rc.1",
     "@carto/react-redux": "^1.0.0-rc.1",
     "@carto/react-ui": "^1.0.0-rc.1",
     "@material-ui/core": "^4.11.3",
